requirements.txt
ruamel.yaml
python-telegram-bot
psutil
